+++ app/Definitions/CreateMode.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Definitions;
+
+/**
+ * The mode used for creating new models
+ *
+ * This is to aid in dynamically selecting methods for finding,
+ * inserting, or updating models.
+ */
+enum CreateMode: string {
+	case FirstOrCreate = 'firstOrCreate';
+	case UpdateOrCreate = 'updateOrCreate';
+}