+++ app/Services/Twitter/TwitterService.php
@@ -6,53 +6,43 @@ use App\Contracts\SocialMediaService;
 use App\DataTransferObjects\TweetDTO;
 use App\DataTransferObjects\TwitterUserDTO;
 use App\Models\Token;
+use App\Services\AbstractEndpoint;
+use App\Services\Twitter\Endpoints\TokenEndpoint;
+use App\Services\Twitter\Endpoints\UsersLookupEndpoint;
+use App\Services\Twitter\Endpoints\UserTimelineEndpoint;
 use Exception;
+use Illuminate\Http\Client\PendingRequest;
+use Illuminate\Http\Client\Response;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Http;
+use Illuminate\Support\Str;
 
 class TwitterService implements SocialMediaService
 {
 	/**
 	 * The base Twitter API URL
-	 *
-	 * @property string         $api_url
-	 * @access private
 	 */
-	private $api_url = 'https://api.twitter.com';
+	private string $api_url = 'https://api.twitter.com';
 
 	/**
 	 * The token used for retrieving tweet information from the Twitter API
-	 *
-	 * @property string         $token
-	 * @access private
 	 */
-	private $token;
+	private ?Token $token;
 
 	/**
 	 * The API key used for generating the token
-	 *
-	 * @property string         $key
-	 * @access private
 	 */
-	private $key;
+	private string $key;
 
 	/**
 	 * The API secret key part used for generating the token
-	 *
-	 * @property string         $secret
-	 * @access private
 	 */
-	private $secret;
+	private string $secret;
 
 	/**
 	 * Create a new instance of the Tweet model
 	 *
-	 * @method                  __construct
-	 * @access public
-	 *
-	 * @param array             $attributes
-	 *
-	 * @return void
+	 * @todo: need to fix 'valid' issue; 'expires_at' is being set to null
 	 *
 	 * This is necessary to initialize some properties that can't otherwise be
 	 * initialized. Initializing properties outside of a constrctor requires
@@ -60,59 +50,51 @@ class TwitterService implements SocialMediaService
 	 */
 	public function __construct()
 	{
-		$this->token = Token::whereRaw("LOWER('service') like '%twitter%'")
+		$this->key = config('services.twitter.key');
+		$this->secret = config('services.twitter.secret');
+
+		$this->token = Token::whereRaw("LOWER(service) like '%twitter%'")
 			->latest()
 			->valid()
-			->first();
-
-		$this->key = config('services.twitter.key', false);
-		$this->secret = config('services.twitter.secret', false);
-
-		if(!$this->key) {
-			throw new Exception("Config option 'services.twitter.key' not set.");
-		}
-
-		if(!$this->secret) {
-			throw new Exception("Config option 'services.twitter.secret' not set.");
-		}
+			->first()
+			?? $this->getToken();
 	}
 
 	/**
-	 * Get the Twitter API token
+	 * Call a specified Twitter API endpoint
 	 *
-	 * Generate a new one if necessary.
-	 *
-	 * @method                  getToken
-	 * @access public
-	 *
-	 * @return App\Models\Token;
-	 *
-	 * The 'if' in this case has to do with whether Twitter already has a token
-	 * for use. If the token hasn't been generated previously or if the previous
-	 * token was revoked, then this function will be used to ask Twitter for a
-	 * new one.
-	 *
-	 * This method will also only be called if the token doesn't already exist
-	 * in the environment.
+	 * @todo: make this more generic or find way to jump between helper
+	 * methods like Http::asForm() or Http::withToken() as needed,
+	 * otherwise this will become messy. The former may require the
+	 * endpoint classes knowing what encoding or other specific
+	 * request format is required besides request headers and params.
 	 */
-	public function getToken()
+	public function call(AbstractEndpoint $endpoint): Response
 	{
-		if ($this->token) {
-			return $this->token;
-		}
+		$endpoint_map = [
+			'asForm' => [
+				TokenEndpoint::class,
+			],
+		];
 
-		if (!$this->key || !$this->secret) {
-			abort(500);
-		}
+		return with(
+			Http::withHeaders($endpoint->headers),
+			fn (PendingRequest $pendingRequest): PendingRequest => match(true) {
+				in_array($endpoint::class, $endpoint_map['asForm']) => $pendingRequest->asForm(),
+				default => $pendingRequest,
+			}
+		)
+		->{Str::lower($endpoint->method->value)}(
+			$endpoint->url(),
+			$endpoint->params
+		);
+	}
 
-		$auth_hash = base64_encode(urlencode($this->key) . ':' . urlencode($this->secret));
-
-		$response = Http::asForm()->withHeaders([
-			'Authorization' => "Basic {$auth_hash}",
-		])->post($this->getUrl("oauth2/token"), [
-			'grant_type' => 'client_credentials',
-		]);
-
+	/**
+	 * Check HTTP responses for errors
+	 */
+	private function checkForErrors(Response $response): void
+	{
 		if ($response->failed()) {
 			$response->throw();
 		}
@@ -123,14 +105,26 @@ class TwitterService implements SocialMediaService
 		) {
 			throw new Exception($response["errors"][0]["message"]);
 		}
+	}
 
-		// $this->token = $response['access_token'];
-		$this->token = Token::create([
+	/**
+	 * Get the Twitter API token from Twitter
+	 *
+	 * This method is intended to be used only if $this->token isn't
+	 * set.
+	 */
+	private function getToken(): Token
+	{
+		$response = $this->call(TokenEndpoint::make()->with([
+			'Authorization' => "Basic " . base64_encode(urlencode($this->key) . ':' . urlencode($this->secret)),
+		]));
+
+		$this->checkForErrors($response);
+
+		return Token::create([
 			'service' => 'twitter',
 			'value' => $response['access_token'],
 		]);
-
-		return $this->token;
 	}
 
 	/**
@@ -152,21 +146,20 @@ class TwitterService implements SocialMediaService
 			throw new Exception("'\$count' value cannot be greater than 3200.");
 		}
 
-		$query = collect([
-			'count' => $count,
-			'include_rts' => $reposts,
-			'screen_name' => $username,
-			'since_id' => $since,
-		])->reject(fn ($value, $key) => is_null($value));
+		$response = $this->call(UserTimelineEndpoint::make()->with(
+			headers: [
+				"Authorization" => "Bearer {$this->token->value}",
+			],
+			params: collect([
+				'count' => $count,
+				'include_rts' => $reposts,
+				'screen_name' => $username,
+				'since_id' => $since,
+			])->reject(fn ($value, $key) => is_null($value))
+			->toArray())
+		);
 
-		$response = Http::withToken($this->getToken()->value)
-			->get(
-				$this->getUrl("1.1/statuses/user_timeline.json"), $query->toArray()
-			);
-
-		if ($response->failed()) {
-			$response->throw();
-		}
+		$this->checkForErrors($response);
 
 		return collect($response->json())
 			->transform(fn ($tweet) => new TweetDTO(
@@ -180,6 +173,30 @@ class TwitterService implements SocialMediaService
 
 	public function getUrl(string $url): string
 	{
+		trigger_error('Method ' . __METHOD__ . ' is deprecated', E_USER_DEPRECATED);
+
 		return "{$this->api_url}/{$url}";
+	}
+
+	public function getUsers(Collection $users): Collection
+	{
+		$response = $this->call(UsersLookupEndpoint::make()->with(
+			headers: [
+				"Authorization" => "Bearer {$this->token->value}",
+			],
+			params: [
+				"screen_name" => $users->implode('screen_name', ','),
+			],
+		));
+
+		$this->checkForErrors($response);
+
+		return collect($response->json())
+			->transform(fn ($user) => new TwitterUserDTO(
+				id: $user['id'],
+				name: $user['name'],
+				screen_name: $user['screen_name'],
+				profile_image_url_https: $user['profile_image_url_https'],
+			));
 	}
 }